teacher_server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dateutil==2.8.2
Pillow==10.1.0

